functions/pubsub-positive-analyzer/requirements.txt
google-cloud-language
slack-sdk